change: introduce lexing/decoding errors on same line
It turns out that pre-PEG Python decodes lazily, meaning it must reach the token with an error to issue the error.
Whereas post-PEG Python tokenizes the entire file all in one go.
The consequence is that it's easier to illustrate these errors if the lexing and decoding errors are on the same line.