change: zero division in first line
I trialed this with some people and it turns out that people don't remember whether it is valid to add a string to an int in Python and they were focusing on the wrong mistake. You can add a string to an int in Java and Java will coerce the int to a String, so this is not an error in Java (or JavaScript). And it's not TECHNICALLY an error in C. It is, however, a bad idea.